requirements.txt
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.33.0
azure-storage-blob==12.14.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
databricks-sdk==0.49.0
Deprecated==1.2.18
docker==7.1.0
exceptiongroup==1.2.2
Faker==37.1.0
fastapi==0.115.12
flake8==7.2.0
Flask==3.1.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mlflow==2.21.2
mlflow-skinny==2.21.2
msrest==0.7.1
mypy-extensions==1.0.0
narwhals==1.33.0
numpy==2.2.4
nvidia-nccl-cu12==2.26.2
oauthlib==3.2.2
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pyflakes==3.3.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.44.0
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
zipp==3.21.0
